型定義ファイル作成
