+++ src/components/atoms/Button.tsx
@@ -4,6 +4,7 @@ type PropType = {
   value: string
   func: () => void
 }
+
 const Button: React.FC<PropType> = ({ value, func }) => {
   return (
     <button type="button" onClick={func}>
